+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Removed hardcoded "perfect score" string in `play_round`.
+
 
 ## [1.0.0] - 2022-05-06
 ### Added